Implementare Authorize

+++ OnlineShop/Controllers/CategoriiController.cs
@@ -9,7 +9,6 @@ namespace OnlineShop.Controllers
 	public class CategoriiController : Controller
 	{
 		private readonly ApplicationDbContext db;
-
 		public CategoriiController(ApplicationDbContext context)
 		{
 			db = context;
@@ -28,17 +27,25 @@ namespace OnlineShop.Controllers
 			return View();
 		}
 
+		[Authorize(Roles = "Administrator")]
 		public ActionResult Show(int id)
 		{
-			Categorie categorie = db.Categorii.Find(id);
+			Categorie? categorie = db.Categorii.Find(id);
+			if (categorie == null)
+			{
+				TempData["message"] = "Categoria cautata nu exista";
+				return RedirectToAction("Index");
+			}
 			return View(categorie);
 		}
 
+		[Authorize(Roles = "Administrator")]
 		public ActionResult New()
 		{
 			return View();
 		}
 
+		[Authorize(Roles = "Administrator")]
 		[HttpPost]
 		public ActionResult New(Categorie categ)
 		{
@@ -52,19 +59,33 @@ namespace OnlineShop.Controllers
 			return View(categ);
 		}
 
+		[Authorize(Roles = "Administrator")]
 		public ActionResult Edit(int id)
 		{
-            Categorie categorie = db.Categorii.Find(id);
+			Categorie? categorie = db.Categorii.Find(id);
+
+			if (categorie == null)
+			{
+				TempData["message"] = "Categoria cautata nu exista";
+				return RedirectToAction("Index");
+			}
 			return View(categorie);
 		}
 
+		[Authorize(Roles = "Administrator")]
 		[HttpPost]
 		public ActionResult Edit(int id, Categorie reqCateg)
 		{
-            Categorie categorie = db.Categorii.Find(id);
+			Categorie? categorie = db.Categorii.Find(id);
+
+			if (categorie == null)
+			{
+				TempData["message"] = "Categoria cautata nu exista";
+				return RedirectToAction("Index");
+			}
 			if (ModelState.IsValid)
 			{
-                categorie.Denumire = reqCateg.Denumire;
+				categorie.Denumire = reqCateg.Denumire;
 				db.SaveChanges();
 				TempData["message"] = "Categoria a fost modificata!";
 				return RedirectToAction("Index");
@@ -72,10 +93,17 @@ namespace OnlineShop.Controllers
 			return View(reqCateg);
 		}
 
+		[Authorize(Roles = "Administrator")]
 		[HttpPost]
 		public ActionResult Delete(int id)
 		{
-            Categorie categorie = db.Categorii.Find(id);
+			Categorie? categorie = db.Categorii.Find(id);
+
+			if (categorie == null)
+			{
+				TempData["message"] = "Categoria cautata nu exista";
+				return RedirectToAction("Index");
+			}
 			db.Categorii.Remove(categorie);
 			TempData["message"] = "Categoria a fost stearsa";
 			db.SaveChanges();
